+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add shared-memory implementation of GPU transforms.
-  They are disabled by default, and can be enabled by passing `gpu_method = :shared_memory`
+- Add alternative implementation of GPU transforms based on shared-memory arrays.
+  This is disabled by default, and can be enabled by passing `gpu_method = :shared_memory`
   (default is `:global_memory`).
 
 ## [v0.5.6](https://github.com/jipolanco/NonuniformFFTs.jl/releases/tag/v0.5.6) - 2024-10-17